refactor: centralize inventory templates
