+++ backend/tests/generateInventory.test.js
@@ -1,4 +1,5 @@
 const generateInventory = require('../src/utils/generateInventory');
+const { raceInventory } = require('../src/data/staticInventoryTemplates');
 
 describe('generateInventory', () => {
   it('selects random items for each category', () => {
@@ -11,7 +12,7 @@ describe('generateInventory', () => {
       { item: 'Меч', code: 'меч', amount: 1, bonus: { strength: 2 } },
       { item: 'Щит', code: 'щит', amount: 1, bonus: { defense: 1 } },
       { item: 'Зілля здоров’я', code: 'зілля_здоров’я', amount: 1, bonus: {} },
-      { item: 'Кістяний талісман', code: 'кістяний_талісман', amount: 1, bonus: { strength: 1 } }
+      { item: raceInventory.orc[0].item, code: 'кістяний_талісман', amount: 1, bonus: raceInventory.orc[0].bonus }
     ]);
   });
 
@@ -25,7 +26,7 @@ describe('generateInventory', () => {
       { item: 'Меч', code: 'меч', amount: 1, bonus: { strength: 2 } },
       { item: 'Шкіряна броня', code: 'шкіряна_броня', amount: 1, bonus: { agility: 1 } },
       { item: 'Зілля здоров’я', code: 'зілля_здоров’я', amount: 1, bonus: {} },
-      { item: 'Кістяний талісман', code: 'кістяний_талісман', amount: 1, bonus: { strength: 1 } }
+      { item: raceInventory.orc[0].item, code: 'кістяний_талісман', amount: 1, bonus: raceInventory.orc[0].bonus }
     ]);
   });
 
@@ -39,7 +40,7 @@ describe('generateInventory', () => {
       { item: 'Довгий лук', code: 'довгий_лук', amount: 1, bonus: { agility: 2 } },
       { item: 'Шкіряна броня', code: 'шкіряна_броня', amount: 1, bonus: { agility: 1 } },
       { item: 'Колчан стріл', code: 'колчан_стріл', amount: 1, bonus: {} },
-      { item: 'Кістяний талісман', code: 'кістяний_талісман', amount: 1, bonus: { strength: 1 } }
+      { item: raceInventory.orc[0].item, code: 'кістяний_талісман', amount: 1, bonus: raceInventory.orc[0].bonus }
     ]);
   });
 
@@ -53,7 +54,7 @@ describe('generateInventory', () => {
       { item: 'Жезл зцілення', code: 'жезл_зцілення', amount: 1, bonus: { intellect: 1 } },
       { item: 'Ряса', code: 'ряса', amount: 1, bonus: { health: 1 } },
       { item: 'Зілля лікування', code: 'зілля_лікування', amount: 1, bonus: {} },
-      { item: 'Кістяний талісман', code: 'кістяний_талісман', amount: 1, bonus: { strength: 1 } }
+      { item: raceInventory.orc[0].item, code: 'кістяний_талісман', amount: 1, bonus: raceInventory.orc[0].bonus }
     ]);
   });
